+++ sebastian/src/lib.rs
@@ -16,6 +16,7 @@ pub mod gitgraph;
 pub mod graphlib;
 pub mod journey;
 pub mod jsmap;
+pub mod kanban;
 pub mod packet;
 pub mod pie;
 pub mod quadrant;